requirements.txt
langchain
openai
chromadb
beautifulsoup4
requests
playwright
python-dotenv
notion-client
